requirements.txt
pygame==2.0.0
wget==3.2
opencv_python==4.2.0.34
numpy==1.16.4
h5py==2.9.0
matplotlib==2.2.3
scipy==1.3.1
Pillow==8.0.1
seaborn==0.11.0
